+++ DataLayers/docs/release-notes.md
@@ -1,6 +1,9 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* Fixed layers defaulting to disabled in 1.17.0.
+
 ## 1.17.0
 Released 08 June 2024 for SMAPI 4.0.7 or later.